added new rpc handler logic
