+++ application/src/tui/components/fps.rs
@@ -70,7 +70,11 @@ impl FpsCounter {
 }
 
 impl Component for FpsCounter {
-    fn update(&mut self, _state: &State, action: Action) -> Result<Option<Action>> {
+    fn update(
+        &mut self,
+        state: &crate::state::StateReadGuard<'_>,
+        action: Action,
+    ) -> Result<Option<Action>> {
         match action {
             Action::Tick => self.app_tick()?,
             Action::Render => self.render_tick()?,
